refactor: BoardType Enum 패키지 변경
