+++ src/main/java/com/airoom/airoom/attach/model/dto/AttachmentRequest.java
@@ -1,6 +1,6 @@
 package com.airoom.airoom.attach.model.dto;
 
-import com.airoom.airoom.board.BoardType;
+import com.airoom.airoom.board.entity.BoardType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
