+++ src/main/java/com/airoom/airoom/attach/model/dto/PresignedUrlRequest.java
@@ -1,7 +1,10 @@
 package com.airoom.airoom.attach.model.dto;
 
-import com.airoom.airoom.board.BoardType;
-import lombok.*;
+import com.airoom.airoom.board.entity.BoardType;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
