+++ src/main/java/com/airoom/airoom/attach/model/repository/AttachmentRepository.java
@@ -1,7 +1,7 @@
 package com.airoom.airoom.attach.model.repository;
 
-import com.airoom.airoom.board.BoardType;
 import com.airoom.airoom.board.entity.Attachment;
+import com.airoom.airoom.board.entity.BoardType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
