+++ src/main/java/com/airoom/airoom/subjectboard/model/service/SubjectBoardService.java
@@ -4,8 +4,8 @@ import com.airoom.airoom.attach.model.dto.AttachmentResponse;
 import com.airoom.airoom.attach.model.repository.AttachmentRepository;
 import com.airoom.airoom.attach.model.service.AttachmentService;
 import com.airoom.airoom.attach.model.service.PresignedUrlService;
-import com.airoom.airoom.board.BoardType;
 import com.airoom.airoom.board.entity.Attachment;
+import com.airoom.airoom.board.entity.BoardType;
 import com.airoom.airoom.board.entity.SubjectBoard;
 import com.airoom.airoom.classroom.entity.Classroom;
 import com.airoom.airoom.classroom.entity.ClassroomTeacher;
@@ -13,8 +13,8 @@ import com.airoom.airoom.classroom.model.repository.ClassroomRepository;
 import com.airoom.airoom.classroom.model.repository.ClassroomTeacherRepository;
 import com.airoom.airoom.member.entity.Member;
 import com.airoom.airoom.member.model.repository.MemberRepository;
-import com.airoom.airoom.subjectboard.model.dto.SubjectBoardRequest;
 import com.airoom.airoom.subjectboard.model.dto.SubjectBoardListResponse;
+import com.airoom.airoom.subjectboard.model.dto.SubjectBoardRequest;
 import com.airoom.airoom.subjectboard.model.dto.SubjectBoardViewResponse;
 import com.airoom.airoom.subjectboard.model.repository.SubjectBoardRepository;
 import com.amazonaws.services.kms.model.NotFoundException;
